requirements.txt
pypdf2
python-dotenv
groq
pytesseract
pillow
